\name{estVarCompCI}
\alias{estVarCompCI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimate confidence intervals for variance components based on normal approximation
}
\description{
Estimate confidence intervals for variance components based on normal approximation. 
Variance components are estimated by the function estVarComp
}
\usage{
estVarCompCI(x, prop = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A null model list returned by nullMixedModel.
}
  \item{prop}{
If prop is TRUE, confidence intervals are based on proportion out of the total vairance. 
}
}
\details{
more to come. 
}
\value{
A matrix with variance estiates (or proportions) and lower and upper values of their 95\% confidence intervals.
}

\author{
Matt Conomos
}
\note{
 To be developed more later. 
}


