\name{lui}
\alias{lui}
\docType{data}
\title{Response diversity and functional redundancy of plant communities under land use intensification across 18 land use intensity gradients}
\description{
This dataset contains the correlation coefficients between land use intensity and response diversity (r.FDis) or functional redundancy (r.nbsp) across 18 land use intensity gradients from nine countries and five biomes.
}
\usage{data(lui)}
\format{
  A data frame with 18 observations on the following 4 variables.
  \describe{
    \item{\code{label}}{the labels for each study}
    \item{\code{r.FDis}}{a numeric vector containing the correlation coefficients between land use intensity and response diversity for each land use intensity gradient}
    \item{\code{r.nbsp}}{a numeric vector containing the correlation coefficients between land use intensity and functional redundancy for each land use intensity gradient}
    \item{\code{n}}{a numeric vector containing the sample sizes used for each study}
  }
}
\source{
Laliberte et al. (in preparation) Land use intensification reduces response diversity in plant communities.
}
\keyword{datasets}
