\encoding{latin1}
\name{metacor-package}
\alias{metacor-package}
\alias{metacor}
\docType{package}
\title{
Meta-analysis of correlation coefficients
}
\description{
Implement the DerSimonian-Laird (DSL) and Olkin-Pratt (OP) meta-analytical approaches with correlation coefficients as effect sizes.
}
\details{
\tabular{ll}{
Package: \tab metacor\cr
Type: \tab Package\cr
Version: \tab 1.0-2\cr
Date: \tab 2011-03-21\cr
License: \tab GLP-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Etienne Laliberté

Maintainer: Etienne Laliberté \email{etiennelaliberte@gmail.com} \url{http://www.elaliberte.info/}
}
\references{
Schulze, R. (2004) Meta-analysis: a comparison of approaches. Hogrefe & Huber, Gottingen, Germany.
}
