% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacor_dual.R
\name{metacor_dual}
\alias{metacor_dual}
\title{Effect Sizes and Imputation for Meta-Analysis of Pre-Post Studies and Pre-Post intervention and control groups studies (metacor_dual)}
\usage{
metacor_dual(
  df,
  digits = NULL,
  add_to_df = TRUE,
  method = "both",
  apply_hedges = TRUE,
  SMD_method = "SMDpre",
  MeanDifferences = FALSE,
  impute_method = "none",
  verbose = TRUE,
  report_imputations = FALSE,
  custom_sd_diff_int = NULL,
  custom_sd_diff_con = NULL,
  single_group = FALSE
)
}
\arguments{
\item{df}{Data frame with the necessary columns for intervention and (optionally) control groups.}

\item{digits}{Number of decimal places to round results (default: NULL).}

\item{add_to_df}{Logical. If TRUE, results are added to the original data frame.}

\item{method}{Method for SDdiff calculation (i.e., 'p_value', 'ci', 'both').}

\item{apply_hedges}{Logical. Apply Hedges' g correction? (default: TRUE)}

\item{SMD_method}{Method for effect size (i.e., 'SMDpre', 'SMDchange', 'ScMDpooled', 'ScMDpre').}

\item{MeanDifferences}{Logical. Calculate mean differences and variances? (default: FALSE)}

\item{impute_method}{Imputation method for missing SDdiff (i.e., 'none', 'direct', 'mean', 'cv').}

\item{verbose}{Logical. Print messages during processing? (default: TRUE)}

\item{report_imputations}{Logical. Generate Word imputation report? (default: FALSE)}

\item{custom_sd_diff_int}{List with elements 'row' and 'value' for manual sd_diff_int values.}

\item{custom_sd_diff_con}{List with elements 'row' and 'value' for manual sd_diff_con values.}

\item{single_group}{Logical. Is the design single-group only? (default: FALSE)}
}
\value{
Data frame with calculated variables. Optionally, a Word report ('imputation_report.docx') is generated.
}
\description{
Calculates effect sizes (i.e., SMDpre, SMDchange, ScMDpooled, ScMDpre) and allows for various imputation methods (i.e., none, cv, direct, mean)
for missing SDdiff and correlation coefficients in pre-post meta-analyses, with or without a control group.
Generates a detailed imputation report in Word format.
}
\examples{
# See vignette("metacor-intro") for a complete example.
}
\references{
Higgins, J. P. T., Thomas, J., Chandler, J., Cumpston, M., Li, T., Page, M. J., & Welch, V. A. (Eds.). (2023). Cochrane handbook for systematic reviews of interventions (Version 6.3). Cochrane. https://training.cochrane.org/handbook
Fu, R., Vandermeer, B.W., Shamliyan, T.A., ONeil, M.E., Yazdi, F., Fox, S.H., & Morton, S.C. (2013). Handling Continuous Outcomes in Quantitative Synthesis. Methods Guide for Comparative Effectiveness Reviews. AHRQ Publication No. 13-EHC103-EF. https://effectivehealthcare.ahrq.gov/reports/final.cfm
}
